requirement.txt
pycryptodome
pyDes
flask
chardet
pyjwt
base65536